Extract the oauth client into separate file
